add: shell script

+++ w1/delete_user.sh
@@ -4,6 +4,13 @@
 # input: user
 # output: user
 
-echo -e "User to be deleted:"
-            read del_user
-            sudo userdel -r $del_user
+echo -n "User to be deleted:"
+read del_user
+sudo userdel -r $del_user
+
+if [[ ! $(sudo cat /etc/passwd | grep "$del_user") ]];
+then
+	echo "user $del_user is not exist!"
+else
+	sudo userdel -r $del_user
+fi
